+++ apps/docs/app/layout.config.tsx
@@ -1,5 +1,5 @@
 import type { DocsLayoutProps } from 'fumadocs-ui/layout';
-import { LayoutTemplateIcon } from 'lucide-react';
+import { BookIcon, LayoutTemplateIcon } from 'lucide-react';
 import { utils } from '@/utils/source';
 import { NavChildren, SidebarBanner, Title } from '@/app/layout.client';
 
@@ -16,6 +16,12 @@ export const layoutOptions: Omit<DocsLayoutProps, 'children'> = {
     banner: <SidebarBanner />,
   },
   links: [
+    {
+      icon: <BookIcon />,
+      text: 'Blog',
+      url: '/blog',
+      active: 'nested-url',
+    },
     {
       text: 'Showcase',
       url: '/showcase',