+++ packages/ui/src/components/sidebar.tsx
@@ -11,8 +11,9 @@ import { useTreeContext } from '@/contexts/tree';
 import { useSidebarCollapse } from '@/contexts/sidebar';
 import { ScrollArea, ScrollViewport } from '@/components/ui/scroll-area';
 import { hasActive, isActive } from '@/utils/shared';
-import type { LinkItem } from '@/layout';
+import type { LinkItemType } from '@/layout';
 import { buttonVariants } from '@/theme/variants';
+import { LinkItem } from '@/components/link-item';
 import {
   Collapsible,
   CollapsibleContent,
@@ -21,7 +22,7 @@ import {
 import { ThemeToggle } from './theme-toggle';
 
 export interface SidebarProps {
-  items: LinkItem[];
+  items: LinkItemType[];
 
   /**
    * Open folders by default if their level is lower or equal to a specific level
@@ -65,14 +66,7 @@ const itemVariants = cva(
 const defaultComponents: Components = {
   Folder: FolderNode,
   Separator: SeparatorNode,
-  Item: ({ item }) => (
-    <BaseItem
-      url={item.url}
-      external={item.external}
-      icon={item.icon}
-      text={item.name}
-    />
-  ),
+  Item: PageNode,
 };
 
 const SidebarContext = createContext<SidebarContext>({
@@ -85,7 +79,8 @@ export function Sidebar({
   components,
   defaultOpenLevel = 1,
   collapsible = true,
-  ...props
+  banner,
+  items,
 }: SidebarProps): React.ReactElement {
   const [open, setOpen] = useSidebarCollapse();
   const alwaysShowFooter = Boolean(footer) || collapsible;
@@ -102,7 +97,7 @@ export function Sidebar({
   };
 
   return (
-    <>
+    <SidebarContext.Provider value={context}>
       {collapsible && !open ? (
         <button
           type="button"
@@ -127,57 +122,58 @@ export function Sidebar({
           'max-md:fixed max-md:inset-0 max-md:z-40 max-md:bg-background/80 max-md:pt-16 max-md:backdrop-blur-md max-md:data-[open=false]:hidden',
         )}
       >
-        <SidebarContext.Provider value={context}>
-          <ViewportContent {...props} />
-          <div
-            className={cn(
-              'flex flex-row items-center gap-2 border-t p-3 md:p-2',
-              !alwaysShowFooter && 'md:hidden',
-            )}
-          >
-            {footer}
-            <ThemeToggle className="md:hidden" />
-            {collapsible ? (
-              <button
-                type="button"
-                aria-label="Trigger Sidebar"
-                className={cn(
-                  buttonVariants({
-                    color: 'ghost',
-                    size: 'icon',
-                    className: 'max-md:hidden ms-auto',
-                  }),
-                )}
-                onClick={onCollapse}
-              >
-                <SidebarIcon />
-              </button>
-            ) : null}
-          </div>
-        </SidebarContext.Provider>
+        <ViewportContent>
+          {banner}
+          {items.length > 0 && (
+            <div className="flex flex-col md:hidden">
+              {items.map((item, i) => (
+                <LinkItem key={i} item={item} on="menu" />
+              ))}
+            </div>
+          )}
+        </ViewportContent>
+        <div
+          className={cn(
+            'flex flex-row items-center gap-2 border-t p-3 md:p-2',
+            !alwaysShowFooter && 'md:hidden',
+          )}
+        >
+          {footer}
+          <ThemeToggle className="md:hidden" />
+          {collapsible ? (
+            <button
+              type="button"
+              aria-label="Trigger Sidebar"
+              className={cn(
+                buttonVariants({
+                  color: 'ghost',
+                  size: 'icon',
+                  className: 'max-md:hidden ms-auto',
+                }),
+              )}
+              onClick={onCollapse}
+            >
+              <SidebarIcon />
+            </button>
+          ) : null}
+        </div>
       </Base.SidebarList>
-    </>
+    </SidebarContext.Provider>
   );
 }
 
 function ViewportContent({
-  items = [],
-  banner,
-}: Pick<SidebarProps, 'items' | 'banner'>): React.ReactElement {
+  children,
+}: {
+  children: React.ReactNode;
+}): React.ReactElement {
   const { root } = useTreeContext();
 
   return (
     <ScrollArea className="flex-1">
       <ScrollViewport>
         <div className="flex flex-col gap-8 pb-10 pt-4 max-md:px-4 md:pr-3 md:pt-10">
-          {banner}
-          {items.length > 0 && (
-            <div className="lg:hidden">
-              {items.map((item) => (
-                <BaseItem key={item.url} {...item} nested />
-              ))}
-            </div>
-          )}
+          {children}
           <NodeList items={root.children} />
         </div>
       </ScrollViewport>
@@ -215,13 +211,13 @@ function NodeList({
   );
 }
 
-function BaseItem({
-  icon,
-  external = false,
-  url,
-  text,
+function PageNode({
+  item: { icon, external = false, url, name },
   nested = false,
-}: Omit<LinkItem, 'type'> & { nested?: boolean }): React.ReactElement {
+}: {
+  item: PageTree.Item;
+  nested?: boolean;
+}): React.ReactElement {
   const pathname = usePathname();
   const active = isActive(url, pathname, nested);
 
@@ -232,7 +228,7 @@ function BaseItem({
       className={cn(itemVariants({ active }))}
     >
       {icon ?? (external ? <ExternalLinkIcon /> : null)}
-      {text}
+      {name}
     </Link>
   );
 }
@@ -269,16 +265,6 @@ function FolderNode({
     [active],
   );
 
-  const content = (
-    <>
-      {icon}
-      {name}
-      <ChevronDown
-        className={cn('ms-auto transition-transform', !extend && '-rotate-90')}
-      />
-    </>
-  );
-
   return (
     <Collapsible open={extend} onOpenChange={setExtend}>
       {index ? (
@@ -287,11 +273,25 @@ function FolderNode({
           href={index.url}
           onClick={onClick}
         >
-          {content}
+          {icon}
+          {name}
+          <ChevronDown
+            className={cn(
+              'ms-auto transition-transform',
+              !extend && '-rotate-90',
+            )}
+          />
         </Link>
       ) : (
         <CollapsibleTrigger className={cn(itemVariants({ active }))}>
-          {content}
+          {icon}
+          {name}
+          <ChevronDown
+            className={cn(
+              'ms-auto transition-transform',
+              !extend && '-rotate-90',
+            )}
+          />
         </CollapsibleTrigger>
       )}
       <CollapsibleContent>