+++ packages/contentlayer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # fumadocs-contentlayer
 
+## 1.1.20
+
+### Patch Changes
+
+- fumadocs-core@11.1.0
+
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [98258b5]
+  - fumadocs-core@11.0.8
+
 ## 1.1.18
 
 ### Patch Changes