+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # next-docs-zeta
 
+## 11.1.0
+
+## 11.0.8
+
+### Patch Changes
+
+- 98258b5: Fix regex problems
+
 ## 11.0.7
 
 ### Patch Changes